+++ user-service/src/main/java/com/niuktok/backend/user/service/UserService.java
@@ -6,5 +6,5 @@ import com.niuktok.backend.user.pojo.vo.UserInfoVO;
 public interface UserService {
     void register(UserRegisterDTO userRegisterDTO);
 
-    UserInfoVO getUserInfo(Integer userID);
+    UserInfoVO getUserInfo(Long userID);
 }
